requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cloudinary==1.25.0
Django==3.2.4
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
Pillow==8.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
